requirements.txt
langchain
langchain_community
pdfminer.six
ebooklib
ollama
python-docx
textract
pandas
openpyxl
python-pptx
python-dotenv